pyproject.toml
[project]
name = "dlthub-playground"
version = "0.1.0"
description = "A playground repository for dltHub (data load tool)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-sql-connector[pyarrow]>=4.0.5",
    "dlt==1.14.1",
    "faker>=37.4.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
]
